+++ src/objects/Email.py
@@ -10,7 +10,3 @@ class Email(Field):
     def __init__(self, value):
         super().__init__(value)
         self.validate()
-    
-    
-
-
