+++ src/objects/Note.py
@@ -1,0 +1,3 @@
+
+class Note:
+    pass
